Rename toast directory again
