+++ packages/frontend/src/redux/reducer/notifications.reducer.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import { ToastMessage } from '$blocks/toast123/Toast';
+import { ToastMessage } from '$blocks/Toast/Toast';
 
 interface NotificationState {
   toastMessages: ToastMessage[];
